+++ backend/requirements.txt
@@ -1,3 +1,6 @@
 yfinance==0.2.40
 fastapi==0.114.2
 fastapi-cli==0.0.5
+fastapi
+uvicorn
+pytz
