+++ backend/server.py
@@ -1,6 +1,7 @@
 from yfinance_functions import process_symbol
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
+from ..server import app
 
 app = FastAPI()
